CNTG/requirements.txt
protobuf>=3.20.2,<6.0.0
PyYAML==6.0.2
google-generativeai==0.8.5
openai>=1.30.0